gemm: check shape compatibility (#65)
* Specify width=32 for floats

Otherwise tests fail with the current hypothesis version, because
numbers are generated that cannot be represented as float32. In the
other direction, using a width of 32 is not an issue, since every
float32 can be represented by a float64 without a loss of accuracy.

* gemm: check that the inner shapes of the matrices line up

`gemm` incorrectly allowed multiplication of two matrices that do not
have the same inner sizes K:

(N, K) x (K, M)

Even worse, this leads to reads of uninitialized memory and
non-deterministic results.