Add cdef nogil sgemm and saxpy functions
The motivation for adding these functions is to be able to use them from
`cdef nogil` code from different BLAS implementations interchangeably.
The signature mirrors CBLAS as closely as possible, with the following
simplifications to `sgemm`:

- Remove the order argument, always assume C-order.
- Since a conjugate transpose is not possible on a float matrix, specify
  transpose using booleans.

We also use 32-bit integer sizes and increments as the lowest-common
denominator of different BLAS versions.